Chore : merge 전처리 작업
