+++ src/main/java/com/ohboon/ohboon/controller/user/PasswordConfirm.java
@@ -1,6 +1,6 @@
 package com.ohboon.ohboon.controller.user;
 
-import com.ohboon.ohboon.dao.UserDao;
+import com.ohboon.ohboon.dao.UserDAO;
 import com.ohboon.ohboon.dto.UserDto;
 import com.ohboon.ohboon.utils.ScriptWriter;
 import jakarta.servlet.ServletException;
@@ -27,7 +27,7 @@ public class PasswordConfirm extends HttpServlet {
         HttpSession session = req.getSession();
         String email = (String) session.getAttribute("sessionEmail");
 
-        UserDao userDao = new UserDao();
+        UserDAO userDao = new UserDAO();
         UserDto userDto = userDao.findUserByEmail(email);
 
         if (userDto != null && BCrypt.checkpw(password, userDto.getUserPW())) {
