+++ src/main/java/com/ohboon/ohboon/controller/user/PasswordSearch.java
@@ -1,6 +1,6 @@
 package com.ohboon.ohboon.controller.user;
 
-import com.ohboon.ohboon.dao.UserDao;
+import com.ohboon.ohboon.dao.UserDAO;
 import com.ohboon.ohboon.utils.ScriptWriter;
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletException;
@@ -23,7 +23,7 @@ public class PasswordSearch extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         String email = req.getParameter("email");
-        UserDao userDao = new UserDao();
+        UserDAO userDao = new UserDAO();
 
         // 방법1) 직접적으로 PW 변경
