+++ src/main/java/com/ohboon/ohboon/controller/user/Signup.java
@@ -1,7 +1,7 @@
 package com.ohboon.ohboon.controller.user;
 
 
-import com.ohboon.ohboon.dao.UserDao;
+import com.ohboon.ohboon.dao.UserDAO;
 import com.ohboon.ohboon.dto.UserDto;
 import com.ohboon.ohboon.utils.ScriptWriter;
 import com.ohboon.ohboon.utils.VerificationCodeGenerator;
@@ -27,7 +27,7 @@ public class Signup extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
-        UserDao userDao = new UserDao();
+        UserDAO userDao = new UserDAO();
 
         // 비밀번호 암호화
         String userPW = req.getParameter("userPW");
