+++ src/main/java/com/ohboon/ohboon/controller/user/UserInfoUpdate.java
@@ -1,7 +1,7 @@
 package com.ohboon.ohboon.controller.user;
 
 
-import com.ohboon.ohboon.dao.UserDao;
+import com.ohboon.ohboon.dao.UserDAO;
 import com.ohboon.ohboon.dto.UserDto;
 import com.ohboon.ohboon.utils.ScriptWriter;
 import jakarta.servlet.RequestDispatcher;
@@ -21,7 +21,7 @@ public class UserInfoUpdate extends HttpServlet {
         HttpSession session = req.getSession();
         String nickname = (String) session.getAttribute("sessionNickname");
 
-        UserDao userDao = new UserDao();
+        UserDAO userDao = new UserDAO();
         UserDto infoUserDto = userDao.infoUser(nickname);
 
         if (infoUserDto == null) {
@@ -50,7 +50,7 @@ public class UserInfoUpdate extends HttpServlet {
         Part profilePart = req.getPart("profile");
         String profile = uploadProfileImage(profilePart, req);
 
-        UserDao userDao = new UserDao();
+        UserDAO userDao = new UserDAO();
         UserDto userDto = userDao.findUserByEmail(email);
 
         if (userDto != null) {
