+++ src/main/java/com/ohboon/ohboon/dao/MatchDAO.java
@@ -1,8 +1,10 @@
 package com.ohboon.ohboon.dao;
 
 import com.ohboon.ohboon.dto.MatchDTO;
+import com.ohboon.ohboon.util.MybatisConnectionFactory;
+
 import org.apache.ibatis.session.SqlSession;
-import util.MybatisConnectionFactory;
+
 
 import java.util.HashMap;
 import java.util.Map;
